requirements.txt
streamlit==1.45.0
tensorflow==2.16.1
numpy>=1.23.5
h5py>=3.10.0
protobuf==3.20.3
Pillow==10.2.0
gdown==5.1.0
altair==4.2.0


 # For image processing in Streamlit
